added basic pagination for products

+++ controllers/admin.js
@@ -2,6 +2,8 @@ const Product = require("../models/product");
 const mongodb = require("mongodb");
 const { validationResult } = require("express-validator");
 
+const fileHelper = require("../util/file");
+
 exports.getAddProduct = (req, res, next) => {
   if (!req.session.isLoggedIn) {
     return res.redirect("/login");
@@ -137,7 +139,10 @@ exports.postEditProduct = (req, res, next) => {
         return redirect("/");
       }
       product.title = title;
-      if (image) product.imageUrl = image.path;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       product.description = description;
       product.price = price;
       return product.save().then((result) => {
@@ -154,7 +159,12 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.params.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((prod) => {
+      if (!prod) return new Error("Prodcut not found");
+      fileHelper.deleteFile(prod.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log("DELETED PRODUCT");
       res.redirect("/admin/products");
